crypto_bot/requirements.txt
# Core exchange library
ccxt==4.2.95

# Telegram bot
python-telegram-bot==20.7

# Technical analysis indicators
pandas-ta==0.3.14b0
pandas==2.1.4
numpy==1.26.3

# HTTP / async
aiohttp==3.9.1
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Database
# SQLite is built into Python — no extra package needed

# Scheduling
apscheduler==3.10.4

# Logging / utilities
colorlog==6.8.0
tabulate==0.9.0

# Date/time helpers
python-dateutil==2.8.2

# Testing (optional but included)
pytest==7.4.4
pytest-asyncio==0.23.3
